+++ plugins/dune-rules/src/qcheck_stm.ml
@@ -9,6 +9,7 @@ type config = {
   dune_output : string option;
   module_prefix : string option;
   submodule : string option;
+  domain : bool;
   fork_timeout : int option;
 }
 
@@ -45,10 +46,11 @@ let libraries =
          ~default:Filename.(basename config.interface_file |> chop_extension))
   in
   let k ppf config =
+    let backend = if config.domain then "domain" else "sequential" in
     pf ppf
-      "libraries@ %aqcheck-stm.stm@ qcheck-stm.sequential@ \
-       qcheck-multicoretests-util@ ortac-runtime-qcheck-stm.sequential"
-      library config
+      "libraries@ %aqcheck-stm.stm@ qcheck-stm.%s@ qcheck-multicoretests-util@ \
+       ortac-runtime-qcheck-stm.%s"
+      library config backend backend
   in
   stanza k
 
@@ -63,6 +65,7 @@ let module_prefix =
   optional_argument "--module-prefix" (fun cfg -> cfg.module_prefix)
 
 let submodule = optional_argument "--submodule" (fun cfg -> cfg.submodule)
+let domain cfg = if cfg.domain then [ (fun ppf _ -> pf ppf "--domain") ] else []
 
 let gen_ortac_rule ppf config =
   let args =
@@ -72,6 +75,7 @@ let gen_ortac_rule ppf config =
     :: dep config_file
     :: quiet
     :: module_prefix config
+    @ domain config
     @ submodule config
   in
   let run ppf = run ppf args in