+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [Dune] Add domain flag to ortac-dune qcheck-stm
+  [\#329](https://github.com/ocaml-gospel/ortac/pull/329)
 - [QCheck-STM] Add basic runtime for domain testing
   [\#328](https://github.com/ocaml-gospel/ortac/pull/328)
 - [QCheck-STM] Refactor runtime to allow for OCaml5-only sub-library